requirements.txt
aiofiles==22.1.0
async-timeout==3.0.1
asyncio~=3.4.3
DateTime==5.0
discord-pretty-help==2.0.0
discord.py==2.1.0
jishaku==2.5.1
pymongo==4.3.3
motor==3.1.1
nest-asyncio==1.5.6